Begin to use Webpack to resolve JS dependencies in Gulp task
Since we are already all set up to use Gulp, and now find ourselves needing to be able to resolve dependencies in JS, we are starting to solve that problem by bringing in a Webpack Gulp plugin. As the team further develops the build process, we may be able to expand our use of Webpack and reduce Gulp usage if the team decides to go that direction.