Modal (#78)
* first stab at including the Text-section as the content of the Full Modal

I couldn't remap all the fields. It seems that the problem is when we try to remap the fields with two levels of @include. 

The "text-section" organism includes the "header-group" molecule. I was able to remap the fields from Full Modal to Text-Section, but I was not able to reach the "Header Group". I'm not sure if this is a technical limitation or if I just wasn't able to figure it out.

* Full Modal fully integrated with Text Section