TC-178: Add third-level-menu-items.
Using macros to iterate over nested menu items. Using an duplicate modified macro for slick nav version of the menu to not include 3rd level on mobile. Slick nav wasn't handling it well and this seemed to be the easiest/only way to override it.

Adds support for icons ahead of link names.